+++ app/pages/Accounts/TransactionView/TransactionView.tsx
@@ -47,7 +47,7 @@ function CopiableListElement({
                     </p>
                 </div>
             }
-            right={<CopyButton value={value} />}
+            right={<CopyButton className={styles.copyButton} value={value} />}
         />
     );
 }