+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 -   It is now possible to set a default account, which will be selected initially when opening the app.
 -   "About" menu link has been added to native menu bar under "Help" for windows/linux. It is already accessible on mac under "Concordium desktop wallet" in the native menu.
 -   When an account is synchronizing transactions, it now shows with a spinning icon in the tab header of the transaction log.
+-   Added flows to make baker transactions from accounts with only one set of credentials possible without going through a multi-signature proposal flow.
 
 ### Changed